src/core/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers

from core.models import Message


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username',)


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = ('text', 'author', 'subject',)


class MessageOutputSerializer(serializers.ModelSerializer):
    author = UserSerializer()

    class Meta:
        model = Message
        fields = ('text', 'author',)
