pyproject.toml
[tool.poetry]
name = "umschool"
version = "0.1.0"
description = ""
authors = ["EKilyas <mailofilyas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
channels-redis = "^3.2.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
